requirements.txt
pypylon
fastapi
"uvicorn[standard]"
numpy
opencv-python
pillow
redis